Make ui-components login exempt, and add a basic test
