+++ assets/test_components.py
@@ -1,0 +1,3 @@
+def test_components_browser(client):
+    response = client.get("/ui-components/")
+    assert response.status_code == 200
